--- src/main/java/com/example/turkcellProject/EmployeeProjectApplication.java
@@ -1,15 +1,0 @@
-package com.example.turkcellProject;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-
-@SpringBootApplication
-@EnableFeignClients
-public class TurkcellProjectApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TurkcellProjectApplication.class, args);
-	}
-
-}
