--- src/main/java/com/example/turkcellProject/model/Department.java
@@ -1,20 +1,0 @@
-package com.example.turkcellProject.model;
-
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-public class Department {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-    private String location;
-
-}
