--- src/main/java/com/example/turkcellProject/repository/DepartmentRepository.java
@@ -1,10 +1,0 @@
-package com.example.turkcellProject.repository;
-
-import com.example.turkcellProject.model.Department;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface DepartmentRepository extends JpaRepository<Department, Long> {
-}
-
