--- src/main/java/com/example/turkcellProject/repository/EmployeeRepository.java
@@ -1,9 +1,0 @@
-package com.example.turkcellProject.repository;
-
-import com.example.turkcellProject.model.Employee;
-import org.springframework.stereotype.Repository;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-@Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Long> {
-}
