+++ src/main/java/com/project/employeeProject/EmployeeProjectApplication.java
@@ -1,0 +1,15 @@
+package com.project.employeeProject;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+@SpringBootApplication
+@EnableFeignClients
+public class EmployeeProjectApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(EmployeeProjectApplication.class, args);
+	}
+
+}
