+++ src/main/java/com/project/employeeProject/constant/ErrorMessage.java
@@ -1,0 +1,9 @@
+package com.project.employeeProject.constant;
+
+public class ErrorMessage {
+
+    public static final String EMPLOYEE_NOT_FOUND="Employee not found with id: ";
+    public static final String DEPARTMENT_NOT_FOUND="Department not found with id: ";
+    public static final String EMPLOYEE_NULL="Employee cannot be null.";
+    public static final String DEPARTMENT_NULL="Department cannot be null.";
+}
