+++ src/main/java/com/project/employeeProject/repository/DepartmentRepository.java
@@ -1,0 +1,10 @@
+package com.project.employeeProject.repository;
+
+import com.project.employeeProject.model.Department;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface DepartmentRepository extends JpaRepository<Department, Long> {
+}
+
