+++ src/main/java/com/project/employeeProject/repository/EmployeeRepository.java
@@ -1,0 +1,9 @@
+package com.project.employeeProject.repository;
+
+import com.project.employeeProject.model.Employee;
+import org.springframework.stereotype.Repository;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+@Repository
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+}
