+++ src/main/java/com/project/employeeProject/service/DepartmentService.java
@@ -1,0 +1,19 @@
+package com.project.employeeProject.service;
+
+
+import com.project.employeeProject.model.Department;
+
+import java.util.List;
+
+public interface DepartmentService {
+
+    List<Department> getAllDepartments();
+
+    Department getDepartmentById(Long id);
+
+    Department addDepartment(Department department);
+
+    Department updateDepartment(Long id, Department departmentDetails);
+
+    void deleteDepartment(Long id);
+}
