--- src/test/java/com/example/turkcellProject/EmployeeProjectApplicationTests.java
@@ -1,9 +1,0 @@
-package com.example.turkcellProject;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TurkcellProjectApplicationTests {
-
-}
