+++ src/test/java/com/project/employeeProject/EmployeeProjectApplicationTests.java
@@ -1,0 +1,9 @@
+package com.project.employeeProject;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TurkcellProjectApplicationTests {
+
+}
